Tutorijal 1 zadatak 4

+++ src/ba/unsa/etf/rpr/Main.java
@@ -90,7 +90,7 @@ public class Main {
                 }
             }
             else if(opcija == 3){
-                int cijena = korpa.dajUkupnuCijenuArtikala(), iznos;
+                double cijena = korpa.dajUkupnuCijenuArtikala(), iznos;
                 System.out.println("Ukupna cijena je " + korpa.dajUkupnuCijenuArtikala() + " KM.");
                 do{
                     System.out.print("Unesite ispravan iznos: ");
